requirements.txt
asgiref==3.5.0
Django==4.0.2
django-heroku==0.3.1
djangorestframework==3.13.1
pytz==2021.1
sqlparse==0.4.2
PyYAML==5.4.1
setuptools>=3.0
gunicorn==20.1.0